+++ src/pages/Home/components/postCreation/index.tsx
